break up check.rs
