+++ crates/analyzer/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod check;
 pub mod error;
 pub mod log;
 pub mod rules;
