+++ crates/pyo3/src/app.rs
@@ -2,11 +2,11 @@ use pyo3::prelude::*;
 use pyo3::{exceptions, PyResult};
 use rayon::prelude::*;
 
-use fapolicy_analyzer::check::trust_status;
 use fapolicy_analyzer::log::Event;
 use fapolicy_app::app::State;
 use fapolicy_app::cfg;
 use fapolicy_app::sys::deploy_app_state;
+use fapolicy_trust::stat::check;
 
 use crate::acl::{PyGroup, PyUser};
 use crate::event::PyEvent;
@@ -56,7 +56,7 @@ impl PySystem {
             .values()
             .par_iter()
             .filter(|r| r.is_system())
-            .map(|r| trust_status(&r.trusted))
+            .map(|r| check(&r.trusted))
             .flatten()
             .map(PyTrust::from)
             .collect()
@@ -80,7 +80,7 @@ impl PySystem {
             .values()
             .par_iter()
             .filter(|r| r.is_ancillary())
-            .map(|r| trust_status(&r.trusted))
+            .map(|r| check(&r.trusted))
             .flatten()
             .map(PyTrust::from)
             .collect()
