+++ crates/pyo3/src/lib.rs
@@ -3,7 +3,6 @@ use pyo3::prelude::*;
 pub mod acl;
 pub mod app;
 pub mod event;
-pub mod syscheck;
 pub mod trust;
 
 #[pymodule]
